finished all main functions

+++ server/src/controllers/noteControllers.ts
@@ -2,12 +2,19 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/noteModel";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefine } from "../utils/assertIsDefine";
 
 /* GET ALL NOTES */
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
+
     // RequestHandler <-- type that determind req, res, next // import { RequestHandler } from "express";
     try {
-        const notes = await NoteModel.find().exec();
+        assertIsDefine(authenticatedUserId);
+
+        const notes = await NoteModel.find({
+            userId: authenticatedUserId,
+        }).exec();
         res.status(200).json(notes);
     } catch (error) {
         next(error); // <-- from Error Handler
@@ -17,7 +24,11 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 /* GET ONE NOTE */
 export const getOneNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
+
     try {
+        assertIsDefine(authenticatedUserId);
+
         // check if note id valid (in case that the format of id is INCORRECT)
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid Note Id");
@@ -26,6 +37,13 @@ export const getOneNote: RequestHandler = async (req, res, next) => {
         // check if note id valid (in case that the format of id is CORRECT)
         if (!note) {
             throw createHttpError(404, "Note not found");
+        }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(
+                401,
+                "You are not allowed to acess this note"
+            );
         }
         res.status(200).json(note);
     } catch (error) {
@@ -46,7 +64,12 @@ export const createNote: RequestHandler<
 > = async (req, res, next) => {
     const title = req.body.title;
     const text = req.body.text;
+
+    const authenticatedUserId = req.session.userId;
+
     try {
+        assertIsDefine(authenticatedUserId);
+
         if (!title) {
             throw createHttpError(400, "Title is needed");
         }
@@ -54,6 +77,7 @@ export const createNote: RequestHandler<
             throw createHttpError(400, "text is needed");
         }
         const newNote = await NoteModel.create({
+            userId: authenticatedUserId,
             title: title,
             text: text,
         });
@@ -80,7 +104,12 @@ export const updateNote: RequestHandler<
     const noteId = req.params.noteId;
     const newTitle = req.body.title;
     const newText = req.body.text;
+
+    const authenticatedUserId = req.session.userId;
+
     try {
+        assertIsDefine(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid Note Id");
         }
@@ -93,6 +122,13 @@ export const updateNote: RequestHandler<
         const note = await NoteModel.findById(noteId).exec();
         if (!note) {
             throw createHttpError(404, "Note not found");
+        }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(
+                401,
+                "You are not allowed to acess this note"
+            );
         }
 
         note.title = newTitle;
@@ -109,7 +145,12 @@ export const updateNote: RequestHandler<
 /* DELETE NOTE */
 export const deleteNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+
+    const authenticatedUserId = req.session.userId;
+
     try {
+        assertIsDefine(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid Note Id");
         }
@@ -117,6 +158,14 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
         if (!note) {
             throw createHttpError(404, "Note not found");
         }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(
+                401,
+                "You are not allowed to acess this note"
+            );
+        }
+
         await note.deleteOne();
 
         res.sendStatus(204); // Important to use sendStatus() instad of status() due that we don't send any json
